Refactorizar: separar movimiento, combate y dibujado de Game.cpp (#61)
* Refactorizar: extraer lógica de movimiento a GameMovement.cpp

* Refactorizar: Separada la lógica de combate en GameCombat.cpp

* Reorganizado el dibujado del juego en GameRender.cpp

* Preparado para la release

---------
