+++ debian/changelog
@@ -1,4 +1,4 @@
-roguebot (1.0.0-1) unstable; urgency=medium
+roguebot (1.1.0-1) unstable; urgency=medium
 
   * Initial Debian package for DCA (Entregable 2).
